solana_program::address_lookup_table

Module state

Structs

AddressLookupTable

LookupTableMeta

Address lookup table metadata

Enums

LookupTableStatus

Activation status of a lookup table

ProgramState

Program account states

Constants

LOOKUP_TABLE_MAX_ADDRESSES

The maximum number of addresses that a lookup table can hold

LOOKUP_TABLE_META_SIZE

The serialized size of lookup table metadata

Functions

estimate_last_valid_slot

The lookup table may be in a deactivating state until the `deactivation_slot`` is no longer “recent”. This function returns a conservative estimate for the last block that the table may be used for lookups. This estimate may be incorrect due to skipped blocks, however, if the current slot is lower than the returned value, the table is guaranteed to still be in the deactivating state.