Struct solana_program::clock::Clock

#[repr(C)]
pub struct Clock {
    pub slot: Slot,
    pub epoch_start_timestamp: UnixTimestamp,
    pub epoch: Epoch,
    pub leader_schedule_epoch: Epoch,
    pub unix_timestamp: UnixTimestamp,
}

A representation of network time.

All members of Clock start from 0 upon network boot.

Fields
slot: Slot

The current Slot.

epoch_start_timestamp: UnixTimestamp

The timestamp of the first Slot in this Epoch.

epoch: Epoch

The current Epoch.

leader_schedule_epoch: Epoch

The future Epoch for which the leader schedule has most recently been calculated.

unix_timestamp: UnixTimestamp

The approximate real world time of the current slot.

This value was originally computed from genesis creation time and network time in slots, incurring a lot of drift. Following activation of the timestamp_correction and timestamp_bounding features it is calculated using a validator timestamp oracle.