Module solana_program::entrypoint_deprecated

@brief Solana Rust-based BPF program entry point supported by the original and now deprecated BPFLoader. For more information see ‘./bpf_loader_deprecated.rs’

Constants

SUCCESS
Programs indicate success with a return value of 0

Functions

deserialize⚠
Deserialize the input arguments

Type Definitions

ProcessInstruction
User implemented function to process an instruction

ProgramResult