solana_program
Module epoch_schedule

Modules

sysvar

Structs

EpochSchedule

Constants

DEFAULT_LEADER_SCHEDULE_SLOT_OFFSET

The default number of slots before an epoch starts to calculate the leader schedule.

DEFAULT_SLOTS_PER_EPOCH

The number of slots per epoch after initial network warmup.

MAX_LEADER_SCHEDULE_EPOCH_OFFSET

The maximum number of slots before an epoch starts to calculate the leader schedule.

MINIMUM_SLOTS_PER_EPOCH

The minimum number of slots per epoch during the warmup period.

Type Aliases

Epoch

The unit of time a given leader schedule is honored.

Slot

The unit of time given to a leader for encoding a block.