solana_program
Module epoch_stake

API for retrieving epoch stake information.

On-chain programs can use this API to retrieve the total stake for the current epoch or the stake for a specific vote account using the sol_get_epoch_stake syscall.

Functions

get_epoch_stake_for_vote_account

Get the current epoch stake for a given vote address.

get_epoch_total_stake

Get the current epoch’s total stake.