Module solana_program::feature

Runtime features.

Runtime features provide a mechanism for features to be simultaneously activated across the network. Since validators may choose when to upgrade, features must remain dormant until a sufficient majority of the network is running a version that would support a given feature.

Feature activation is accomplished by:

Activation is requested by the feature authority, who issues a transaction to create the feature account. The newly created feature account will have the value of Feature::default()

When the next epoch is entered the runtime will check for new activation requests and active them. When this occurs, the activation slot is recorded in the feature account

Structs

Feature

__FieldVisitor 🔒

__Visitor 🔒

Enums

__Field 🔒

Constants

FIELDS 🔒

ID

The const program ID.

Functions

activate

Activate a feature

activate_with_lamports

check_id

Returns true if given pubkey is the program ID.

id

Returns the program ID.