solana_program
Module instruction

Structs

AccountMeta

Describes a single account read or written by a program during instruction execution.

CompiledInstruction

A compact encoding of an instruction.

Instruction

A directive for a single invocation of a Solana program.

ProcessedSiblingInstruction

Use to query and convey information about the sibling instruction components when calling the sol_get_processed_sibling_instruction syscall.

Enums

InstructionError

Reasons the runtime might have rejected an instruction.

Constants

TRANSACTION_LEVEL_STACK_HEIGHT

Functions

get_processed_sibling_instruction

Returns a sibling instruction from the processed sibling instruction list.

get_stack_height

Get the current stack height, transaction-level instructions are height TRANSACTION_LEVEL_STACK_HEIGHT, fist invoked inner instruction is height TRANSACTION_LEVEL_STACK_HEIGHT + 1, etc…