solana_program

Module loader_v4

Structs

LoaderV4State

LoaderV4 account states

Enums

LoaderV4Status

Constants

DEPLOYMENT_COOLDOWN_IN_SLOTS

Cooldown before a program can be un-/redeployed again

ID

The const program ID.

Functions

check_id

Returns true if given pubkey is the program ID.

create_buffer

Returns the instructions required to initialize a program/buffer account.

deploy

Returns the instructions required to deploy a program.

deploy_from_source

Returns the instructions required to deploy a program using a buffer.

finalize

Returns the instructions required to finalize program.

id

Returns the program ID.

is_deploy_instruction

is_finalize_instruction

is_retract_instruction

is_set_program_length_instruction

is_transfer_authority_instruction

is_truncate_instruction

is_write_instruction

retract

Returns the instructions required to retract a program.

set_program_length

Returns the instructions required to set the length of the program account.

transfer_authority

Returns the instructions required to transfer authority over a program.

truncate

Returns the instructions required to set the length of the program account.

truncate_uninitialized

Returns the instructions required to set the length of the program account.

write

Returns the instructions required to write a chunk of program data to a buffer account.