solana_program

Macro custom_heap_default

macro_rules! custom_heap_default {
    () => { ... };
}

Define the default global allocator.

The default global allocator is enabled only if the calling crate has not disabled it using Cargo features as described below. It is only defined for BPF targets.

Cargo features

A crate that calls this macro can provide its own custom heap implementation, or allow others to provide their own custom heap implementation, by adding a custom-heap feature to its Cargo.toml. After enabling the feature, one may define their own global allocator in the standard way.