solana_program
Macro entrypoint_no_alloc

macro_rules! entrypoint_no_alloc {
    ($process_instruction:ident) => { ... };
}

Declare the program entrypoint and set up global handlers.

This is similar to the entrypoint! macro, except that it does not perform any dynamic allocations, and instead writes the input accounts into a pre- allocated array.

This version reduces compute unit usage by 20-30 compute units per unique account in the instruction. It may become the default option in a future release.

For more information about how the program entrypoint behaves and what it does, please see the documentation for entrypoint!.

NOTE: This entrypoint has a hard-coded limit of 64 input accounts.