Enum solana_program::message::SanitizedMessage

pub enum SanitizedMessage {
    Legacy(LegacyMessage),
    V0(LoadedMessage<'static>),
}

Sanitized message of a transaction.

Variants

Legacy(LegacyMessage)

Sanitized legacy message

V0(LoadedMessage<'static>)

Sanitized version #0 message with dynamically loaded addresses

Implementations

impl SanitizedMessage

pub fn has_duplicates(&self) -> bool

Return true if this message contains duplicate account keys

pub fn header(&self) -> &MessageHeader

Message header which identifies the number of signer and writable or readonly accounts

pub fn legacy_message(&self) -> Option<&LegacyMessage>

Returns a legacy message if this sanitized message wraps one

pub fn fee_payer(&self) -> &Pubkey

Returns the fee payer for the transaction

pub fn recent_blockhash(&self) -> &Hash

The hash of a recent block, used for timing out a transaction

pub fn instructions(&self) -> &[CompiledInstruction]

Program instructions that will be executed in sequence and committed in one atomic transaction if all succeed.

pub fn program_instructions_iter(
&self
) -> impl Iterator<Item = (&Pubkey, &CompiledInstruction)>

Program instructions iterator which includes each instruction’s program id.

pub fn account_keys(&self) -> AccountKeys<'_>

Returns the list of account keys that are loaded for this message.

pub fn is_invoked(&self, key_index: usize) -> bool

Returns true if the account at the specified index is invoked as a program in this message.

pub fn is_non_loader_key(&self, key_index: usize) -> bool

Returns true if the account at the specified index is not invoked as a program or, if invoked, is passed to a program.

pub fn is_writable(&self, index: usize) -> bool

Returns true if the account at the specified index is writable by the instructions in this message.

pub fn is_signer(&self, index: usize) -> bool

Returns true if the account at the specified index signed this message.

pub fn num_readonly_accounts(&self) -> usize

Return the number of readonly accounts loaded by this message.

pub fn decompile_instructions(&self) -> Vec<BorrowedInstruction<'_>>

Decompile message instructions without cloning account keys

pub fn is_upgradeable_loader_present(&self) -> bool

Inspect all message keys for the bpf upgradeable loader

pub fn get_ix_signers(&self, ix_index: usize) -> impl Iterator<Item = &Pubkey>

Get a list of signers for the instruction at the given index

pub fn get_durable_nonce(&self, nonce_must_be_writable: bool) -> Option<&Pubkey>

If the message uses a durable nonce, return the pubkey of the nonce account