Module solana_program::message::legacy

The original and current Solana message format.

This crate defines two versions of Message in their own modules: legacy and v0. legacy is the current version as of Solana 1.10.0. v0 is a future message format that encodes more account keys into a transaction than the legacy format.

Structs

BUILTIN_PROGRAMS_KEYS

Message

A Solana transaction message (legacy).