Module solana_program::slot_hashes

named accounts for synthesized data accounts for bank state, etc.

this account carries the Bank’s most recent bank hashes for some N parents

Re-exports

pub use crate::clock::Slot;

Structs

SlotHashes

Constants

MAX_ENTRIES

Functions

get_entries

set_entries_for_tests_only

Type Definitions

SlotHash