solana_program
Module stake

Modules

config

Config for staking.

instruction

program

stake_flags

state

tools

Utility functions

Constants

MINIMUM_DELINQUENT_EPOCHS_FOR_DEACTIVATION

The minimum number of epochs before stake account that is delegated to a delinquent vote account may be unstaked with StakeInstruction::DeactivateDelinquent