solana_program::stake
Module config

Config for staking.

It carries variables that the stake program cares about.

Structs

ConfigDeprecated

Constants

DEFAULT_SLASH_PENALTY

DEFAULT_WARMUP_COOLDOWN_RATE

ID

The const program ID.

Functions

check_idDeprecated

Returns true if given pubkey is the program ID.

idDeprecated

Returns the program ID.