solana_program::system_instruction

Enum SystemInstruction

pub enum SystemInstruction {

    CreateAccount {
        lamports: u64,
        space: u64,
        owner: Pubkey,
    },
    Assign {
        owner: Pubkey,
    },
    Transfer {
        lamports: u64,
    },
    CreateAccountWithSeed {
        base: Pubkey,
        seed: String,
        lamports: u64,
        space: u64,
        owner: Pubkey,
    },
    AdvanceNonceAccount,
    WithdrawNonceAccount(u64),
    InitializeNonceAccount(Pubkey),
    AuthorizeNonceAccount(Pubkey),
    Allocate {
        space: u64,
    },
    AllocateWithSeed {
        base: Pubkey,
        seed: String,
        space: u64,
        owner: Pubkey,
    },
    AssignWithSeed {
        base: Pubkey,
        seed: String,
        owner: Pubkey,
    },
    TransferWithSeed {
        lamports: u64,
        from_seed: String,
        from_owner: Pubkey,
    },
    UpgradeNonceAccount,

}

An instruction to the system program.

Variants

CreateAccount

Create a new account

Account references

[WRITE, SIGNER] Funding account

[WRITE, SIGNER] New account

Fields

lamports: u64

Number of lamports to transfer to the new account

space: u64

Number of bytes of memory to allocate

owner: Pubkey

Address of program that will own the new account

Assign

Assign account to a program

Account references

[WRITE, SIGNER] Assigned account public key

Fields

owner: Pubkey

Owner program account

Transfer

Transfer lamports

Account references

[WRITE, SIGNER] Funding account

[WRITE] Recipient account

Fields

lamports: u64

CreateAccountWithSeed

Create a new account at an address derived from a base pubkey and a seed

Account references

[WRITE, SIGNER] Funding account

[WRITE] Created account

[SIGNER] (optional) Base account; the account matching the base Pubkey below must be provided as a signer, but may be the same as the funding account and provided as account 0

Fields

base: Pubkey

Base public key

seed: String

String of ASCII chars, no longer than Pubkey::MAX_SEED_LEN

lamports: u64

Number of lamports to transfer to the new account

space: u64

Number of bytes of memory to allocate

owner: Pubkey

Owner program account address

AdvanceNonceAccount

Consumes a stored nonce, replacing it with a successor

Account references

[WRITE] Nonce account

[] RecentBlockhashes sysvar

[SIGNER] Nonce authority

WithdrawNonceAccount(u64)

Withdraw funds from a nonce account

Account references

[WRITE] Nonce account

[WRITE] Recipient account

[] RecentBlockhashes sysvar

[] Rent sysvar

[SIGNER] Nonce authority

The u64 parameter is the lamports to withdraw, which must leave the account balance above the rent exempt reserve or at zero.

InitializeNonceAccount(Pubkey)

Drive state of Uninitialized nonce account to Initialized, setting the nonce value

Account references

[WRITE] Nonce account

[] RecentBlockhashes sysvar

[] Rent sysvar

The Pubkey parameter specifies the entity authorized to execute nonce instruction on the account

No signatures are required to execute this instruction, enabling derived nonce account addresses

AuthorizeNonceAccount(Pubkey)

Change the entity authorized to execute nonce instructions on the account

Account references

[WRITE] Nonce account

[SIGNER] Nonce authority

The Pubkey parameter identifies the entity to authorize

Allocate

Allocate space in a (possibly new) account without funding

Account references

[WRITE, SIGNER] New account

Fields

space: u64

Number of bytes of memory to allocate

AllocateWithSeed

Allocate space for and assign an account at an address derived from a base public key and a seed

Account references

[WRITE] Allocated account

[SIGNER] Base account

Fields

base: Pubkey

Base public key

seed: String

String of ASCII chars, no longer than pubkey::MAX_SEED_LEN

space: u64

Number of bytes of memory to allocate

owner: Pubkey

Owner program account

AssignWithSeed

Assign account to a program based on a seed

Account references

[WRITE] Assigned account

[SIGNER] Base account

Fields

base: Pubkey

Base public key

seed: String

String of ASCII chars, no longer than pubkey::MAX_SEED_LEN

owner: Pubkey

Owner program account

TransferWithSeed

Transfer lamports from a derived address

Account references

[WRITE] Funding account

[SIGNER] Base for funding account

[WRITE] Recipient account

Fields

lamports: u64

Amount to transfer

from_seed: String

Seed to use to derive the funding account address

from_owner: Pubkey

Owner to use to derive the funding account address

UpgradeNonceAccount

One-time idempotent upgrade of legacy nonce versions in order to bump them out of chain blockhash domain.

Account references

[WRITE] Nonce account