solana_program

Module sysvar

Modules

clock

Information about the network’s clock, ticks, slots, etc.

epoch_rewards

Epoch rewards for current epoch

epoch_schedule

Information about epoch duration.

fees

Current cluster fees.

instructions

last_restart_slot

Information about the last restart slot (hard fork).

recent_blockhashes

Information about recent blocks and their fee calculators.

rent

Configuration for network rent.

rewards

This sysvar is deprecated and unused.

slot_hashes

The most recent hashes of a slot’s parent banks.

slot_history

A bitvector of slots present over the last epoch.

Macros

declare_deprecated_sysvar_id

Same as declare_sysvar_id except that it reports that this ID has been deprecated.

declare_sysvar_id

Declares an ID that implements SysvarId.

Constants

ID

The const program ID.

Traits

Sysvar

Interface for loading a sysvar.

SysvarId

A type that holds sysvar data and has an associated sysvar Address.

SysvarSerialize

A type that holds sysvar data.

Functions

check_id

Returns true if given address is the ID.

id

Returns the ID.