solana_program::sysvar

Module instructions

The serialized instructions of the current transaction.

The instructions sysvar provides access to the serialized instruction data for the currently-running transaction. This allows for instruction introspection, which is required for correctly interoperating with native programs like the secp256k1 and ed25519 programs.

Unlike other sysvars, the data in the instructions sysvar is not accessed through a type that implements the Sysvar trait. Instead, the instruction sysvar is accessed through several free functions within this module.

See also the Solana documentation on the instructions sysvar.

Examples

For a complete example of how the instructions sysvar is used see the documentation for secp256k1_instruction in the solana-sdk crate.

Structs

BorrowedAccountMeta

Borrowed version of AccountMeta.

BorrowedInstruction

Borrowed version of Instruction.

Instructions

Instructions sysvar, dummy type.

Constants

ID

The const program ID.

Functions

check_id

Returns true if given pubkey is the program ID.

construct_instructions_data

Construct the account data for the instructions sysvar.

get_instruction_relative

Returns the Instruction relative to the current Instruction in the currently executing Transaction.

id

Returns the program ID.

load_current_indexDeprecated

Load the current Instruction’s index in the currently executing Transaction.

load_current_index_checked

Load the current Instruction’s index in the currently executing Transaction.

load_instruction_atDeprecated

Load an Instruction in the currently executing Transaction at the specified index.

load_instruction_at_checked

Load an Instruction in the currently executing Transaction at the specified index.

store_current_index

Store the current Instruction’s index in the instructions sysvar data.