Module solana_program::sysvar::recent_blockhashes

Information about recent blocks and their fee calculators.

The recent blockhashes sysvar provides access to the RecentBlockhashes, which contains recent blockhahes and their FeeCalculators.

RecentBlockhashes does not implement Sysvar::get.

This sysvar is deprecated and should not be used. Transaction fees should be determined with the getFeeForMessage RPC method. For additional context see the Comprehensive Compute Fees proposal.

See also the Solana documentation on the recent blockhashes sysvar.

Structs

EntryDeprecated

IntoIterSorted

IterItemDeprecated

RecentBlockhashesDeprecated

Contains recent block hashes and fee calculators.

__FieldVisitor 🔒

__Visitor 🔒

__Visitor 🔒

Enums

__Field 🔒

Constants

FIELDS 🔒

MAX_ENTRIESDeprecated

Statics

ID

The static program ID.

Functions

check_idDeprecated

Returns true if given pubkey is the program ID.

idDeprecated

Returns the program ID.