solana_program::sysvar::slot_history

Struct AccountInfo

#[repr(C)]
pub struct AccountInfo<'a> {
    pub key: &'a Pubkey,
    pub lamports: Rc<RefCell<&'a mut u64>>,
    pub data: Rc<RefCell<&'a mut [u8]>>,
    pub owner: &'a Pubkey,
    pub rent_epoch: u64,
    pub is_signer: bool,
    pub is_writable: bool,
    pub executable: bool,
}

Account information

Fields
key: &'a Pubkey

Public key of the account

lamports: Rc<RefCell<&'a mut u64>>

The lamports in the account. Modifiable by programs.

data: Rc<RefCell<&'a mut [u8]>>

The data held in this account. Modifiable by programs.

owner: &'a Pubkey

Program that owns this account

rent_epoch: u64

The epoch at which this account will next owe rent

is_signer: bool

Was the transaction signed by this accountโs public key?

is_writable: bool

Is the account writable?

executable: bool

This accountโs data contains a loaded program (and is now read-only)

Implementations

impl<'a> AccountInfo<'a>

pub fn signer_key(&self) -> Option<&Pubkey>

pub fn unsigned_key(&self) -> &Pubkey

pub fn lamports(&self) -> u64

pub fn try_lamports(&self) -> Result<u64, ProgramError>

pub unsafe fn original_data_len(&self) -> usize

Return the accountโs original data length when it was serialized for the current program invocation.

Safety

This method assumes that the original data length was serialized as a u32 integer in the 4 bytes immediately preceding the serialized account key.

pub fn data_len(&self) -> usize

pub fn try_data_len(&self) -> Result<usize, ProgramError>

pub fn data_is_empty(&self) -> bool

pub fn try_data_is_empty(&self) -> Result<bool, ProgramError>

pub fn try_borrow_lamports(&self) -> Result<Ref<'_, &mut u64>, ProgramError>

pub fn try_borrow_mut_lamports( &self, ) -> Result<RefMut<'_, &'a mut u64>, ProgramError>

pub fn try_borrow_data(&self) -> Result<Ref<'_, &mut [u8]>, ProgramError>

pub fn try_borrow_mut_data( &self, ) -> Result<RefMut<'_, &'a mut [u8]>, ProgramError>

pub fn realloc( &self, new_len: usize, zero_init: bool, ) -> Result<(), ProgramError>

Realloc the accountโs data and optionally zero-initialize the new memory.

Note: Account data can be increased within a single call by up to solana_program::entrypoint::MAX_PERMITTED_DATA_INCREASE bytes.

Note: Memory used to grow is already zero-initialized upon program entrypoint and re-zeroing it wastes compute units. If within the same call a program reallocs from larger to smaller and back to larger again the new space could contain stale data. Pass true for zero_init in this case, otherwise compute units will be wasted re-zero-initializing.

Safety

This method makes assumptions about the layout and location of memory referenced by AccountInfo fields. It should only be called for instances of AccountInfo that were created by the runtime and received in the process_instruction entrypoint of a program.

pub fn assign(&self, new_owner: &Pubkey)

pub fn new( key: &'a Pubkey, is_signer: bool, is_writable: bool, lamports: &'a mut u64, data: &'a mut [u8], owner: &'a Pubkey, executable: bool, rent_epoch: u64, ) -> AccountInfo<'a>

pub fn deserialize_data<T>(&self) -> Result<T, Box<ErrorKind>>
where T: DeserializeOwned,

pub fn serialize_data<T>(&self, state: &T) -> Result<(), Box<ErrorKind>>
where T: Serialize,