Crate solana_program

Modules

account_info

borsh
Borsh utils

bpf_loader
@brief The latest Solana BPF loader.

bpf_loader_deprecated
@brief The original and now deprecated Solana BPF loader.

bpf_loader_upgradeable
@brief An Upgradeable Solana BPF loader.

clock
Provides information about the network’s clock which is made up of ticks, slots, etc…

config

decode_error

entrypoint
@brief Solana Rust-based BPF program entry point supported by the latest BPFLoader. For more information see ‘./bpf_loader.rs’

entrypoint_deprecated
@brief Solana Rust-based BPF program entry point supported by the original and now deprecated BPFLoader. For more information see ‘./bpf_loader_deprecated.rs’

epoch_schedule
configuration for epochs, slots

feature

fee_calculator

hash
The hash module provides functions for creating SHA-256 hashes.

incinerator
Lamports credited to this address will be removed from the total supply (burned) at the end of the current block.

instruction
Defines a composable Instruction type and a memory-efficient CompiledInstruction.

keccak

lamports

loader_instruction

loader_upgradeable_instruction
Upgradeable loader instruction definitions

log
@brief Solana Rust-based BPF program logging

message
A library for generating a message from a sequence of instructions

native_token

nonce

program

program_error

program_memory
@brief Solana Rust-based BPF memory operations

program_option
A C representation of Rust’s std::option::Option used accross the FFI boundary for Solana program interfaces

program_pack
State transition types

program_stubs
@brief Syscall stubs when building for programs for non-BPF targets

pubkey

rent
configuration for network rent

sanitize

secp256k1_program

secp256k1_recover

serialize_utils

short_vec

slot_hashes
named accounts for synthesized data accounts for bank state, etc.

slot_history
slot history

stake

stake_history
named accounts for synthesized data accounts for bank state, etc.

system_instruction

system_program

sysvar
named accounts for synthesized data accounts for bank state, etc.

vote

Macros

custom_heap_default
Fallback to default for unused custom heap feature.

custom_panic_default
Fallback to default for unused custom panic feature. This must be used if the entrypoint! macro is not used.

declare_id
Convenience macro to declare a static public key and functions to interact with it

declare_sysvar_id

entrypoint
Declare the entry point of the program and use the default local heap implementation

entrypoint_deprecated
Declare the entry point of the program.

impl_sysvar_get

info	Deprecated

msg
Print a message to the log

unchecked_div_by_const
Convenience macro for doing integer division where the opersation’s safety can be checked at compile-time