Module solana_program::bpf_loader_upgradeable

@brief An Upgradeable Solana BPF loader.

The upgradeable BPF loader is responsible for deploying, upgrading, and executing BPF programs. The upgradeable loader allows a program’s authority to update the program at any time. This ability break’s the “code is law” contract the usually enforces the policy that once a program is on-chain it becomes immutable. Because of this, care should be taken before executing upgradeable programs which still have a functioning authority. For more information refer to loader_upgradeable_instruction.rs

Enums

UpgradeableLoaderState
Upgradeable loader account states

Statics

ID
The static program ID

Functions

check_id
Confirms that a given pubkey is equivalent to the program ID

close
Returns the instructions required to close a buffer account

close_any
Returns the instructions required to close program, buffer, or uninitialized account

create_buffer
Returns the instructions required to initialize a Buffer account.

deploy_with_max_program_len
Returns the instructions required to deploy a program with a specified maximum program length. The maximum length must be large enough to accommodate any future upgrades.

id
Returns the program ID

is_close_instruction

is_set_authority_instruction

is_upgrade_instruction

set_buffer_authority
Returns the instructions required to set a buffers’s authority.

set_upgrade_authority
Returns the instructions required to set a program’s authority.

upgrade
Returns the instructions required to upgrade a program.

write
Returns the instructions required to write a chunk of program data to a buffer account.