Module solana_program::entrypoint

The Rust-based BPF program entrypoint supported by the latest BPF loader.

For more information see the bpf_loader module.

Structs

BumpAllocator

The bump allocator used as the default rust heap when running programs.

Constants

BPF_ALIGN_OF_U128

assert_eq(std::mem::align_of::<u128>(), 8) is true for BPF but not for some host machines

HEAP_LENGTH

Length of the heap memory region used for program heap.

HEAP_START_ADDRESS

Start address of the memory region used for program heap.

MAX_PERMITTED_DATA_INCREASE

Maximum number of bytes a program may add to an account during a single realloc

NON_DUP_MARKER

Value used to indicate that a serialized account is not a duplicate

SUCCESS

Programs indicate success with a return value of 0

Functions

deserialize⚠

Deserialize the input arguments

Type Definitions

ProcessInstruction

User implemented function to process an instruction

ProgramResult