Module solana_program::epoch_schedule

configuration for epochs, slots

Re-exports

pub use crate::clock::Epoch;

pub use crate::clock::Slot;

pub use crate::clock::DEFAULT_SLOTS_PER_EPOCH;

Structs

EpochSchedule

Constants

DEFAULT_LEADER_SCHEDULE_SLOT_OFFSET
The number of slots before an epoch starts to calculate the leader schedule. Default is an entire epoch, i.e. leader schedule for epoch X is calculated at the beginning of epoch X - 1.

MAX_LEADER_SCHEDULE_EPOCH_OFFSET
The maximum number of slots before an epoch starts to calculate the leader schedule. Default is an entire epoch, i.e. leader schedule for epoch X is calculated at the beginning of epoch X - 1.

MINIMUM_SLOTS_PER_EPOCH
based on MAX_LOCKOUT_HISTORY from vote_program