Module solana_program::message::legacy

The original and current Solana message format.

This crate defines two versions of Message in their own modules: legacy and v0. legacy is the current version as of Solana 1.10.0. v0 is a future message format that encodes more account keys into a transaction than the legacy format.

Structs

BUILTIN_PROGRAMS_KEYS

MAYBE_BUILTIN_KEY_OR_SYSVAR

Message

A Solana transaction message (legacy).

__DeserializeWith ğŸ”’

__DeserializeWith ğŸ”’

__DeserializeWith ğŸ”’

__DeserializeWith ğŸ”’

__FieldVisitor ğŸ”’

__SerializeWith ğŸ”’

__SerializeWith ğŸ”’

__Visitor ğŸ”’

Enums

__Field ğŸ”’

Constants

FIELDS ğŸ”’

Functions

__stability ğŸ”’

__stability ğŸ”’

__static_ref_initialize ğŸ”’

__static_ref_initialize ğŸ”’

__wbg_message_new ğŸ”’ âš

compile_instruction ğŸ”’

compile_instructions ğŸ”’

is_builtin_key_or_sysvar

position ğŸ”’