Trait solana_program::sanitize::Sanitize

pub trait Sanitize {
    fn sanitize(&self) -> Result<(), SanitizeError> { ... }
}

Trait for sanitizing values and members of over the wire messages. Implementation should recursively decent through the data structure and sanitize all struct members and enum clauses. Sanitize excludes signature verification checks, those are handled by another pass. Sanitize checks should include but are not limited too:

All index values are in range

All values are within their static max/min bounds

Provided methods

fn sanitize(&self) -> Result<(), SanitizeError>