Module solana_program::system_instruction

Enums

NonceError

SystemError

SystemInstruction

Constants

MAX_PERMITTED_DATA_LENGTH

Maximum permitted size of data: 10 MiB

Functions

advance_nonce_account

allocate

allocate_with_seed

assign

assign_with_seed

authorize_nonce_account

create_account

create_account_with_seed

create_nonce_account

create_nonce_account_with_seed

instruction_to_nonce_error

nonce_to_instruction_error

transfer

transfer_many

Create and sign new SystemInstruction::Transfer transaction to many destinations

transfer_with_seed

upgrade_nonce_account

One-time idempotent upgrade of legacy nonce versions in order to bump them out of chain blockhash domain.

withdraw_nonce_account