solana_program

Module system_instruction

Enums

SystemError

SystemInstruction

An instruction to the system program.

Constants

MAX_PERMITTED_ACCOUNTS_DATA_ALLOCATIONS_PER_TRANSACTION

Maximum permitted size of new allocations per transaction, in bytes.

MAX_PERMITTED_DATA_LENGTH

Maximum permitted size of account data (10 MiB).

Functions

advance_nonce_account

Advance the value of a durable transaction nonce.

allocate

Allocate space for an account.

allocate_with_seed

assign

Assign ownership of an account from the system program.

assign_with_seed

authorize_nonce_account

Change the authority of a durable transaction nonce account.

create_account

Create an account.

create_account_with_seed

create_nonce_account

Create an account containing a durable transaction nonce.

create_nonce_account_with_seed

transfer

Transfer lamports from an account owned by the system program.

transfer_many

Transfer lamports from an account owned by the system program to multiple accounts.

transfer_with_seed

upgrade_nonce_account

One-time idempotent upgrade of legacy nonce versions in order to bump them out of chain blockhash domain.

withdraw_nonce_account

Withdraw lamports from a durable transaction nonce account.