Module solana_program::sysvar

named accounts for synthesized data accounts for bank state, etc.

Modules

clock
This account contains the clock slot, epoch, and leader_schedule_epoch

epoch_schedule
This account contains the current cluster rent

fees
This account contains the current cluster fees

instructions
This account contains the serialized transaction instructions

recent_blockhashes

rent
This account contains the current cluster rent

rewards
DEPRECATED: This sysvar can be removed once the pico-inflation feature is enabled

slot_hashes
named accounts for synthesized data accounts for bank state, etc.

slot_history
named accounts for synthesized data accounts for bank state, etc.

stake_history
named accounts for synthesized data accounts for bank state, etc.

Statics

ID
The static program ID

Traits

Sysvar

SysvarId

Functions

check_id
Confirms that a given pubkey is equivalent to the program ID

id
Returns the program ID

is_sysvar_id