Module solana_program::sysvar::instructions

This account contains the serialized transaction instructions

Structs

BorrowedAccountMeta

Borrowed version of AccountMeta

BorrowedInstruction

Borrowed version of Instruction

Instructions

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

construct_instructions_data

get_instruction_relative

Returns the Instruction relative to the current Instruction in the currently executing Transaction

id

Returns the program ID

load_current_indexDeprecated

Load the current Instruction’s index in the currently executing Transaction from the Instructions Sysvar data

load_current_index_checked

Load the current Instruction’s index in the currently executing Transaction

load_instruction_atDeprecated

Load an Instruction in the currently executing Transaction at the specified index

load_instruction_at_checked

Load an Instruction in the currently executing Transaction at the specified index

store_current_index

Store the current Instruction’s index in the Instructions Sysvar data